todo-service/requirements.txt
flask
flask_sqlalchemy
python-dotenv
pytest
pytest-flask
gunicorn
pymysql
cryptography>=41.0.0